Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,9 +1,9 @@
 // Your code here...
 #include <stdio.h>
 int main(){
-    int a;
-    int b;
-    int c;
+    float a;
+    float b;
+    float c;
     float avg;
     scanf("%d %d %d", &a,&b,&c);
     avg = (a+b+c)/3;
